@@ -1,6 +1,6 @@
 {
   "name": "@rxstack/sequelize-service",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "RxStack Sequelize Service",
   "private": false,
   "author": "Nikolay Georgiev <symfonist@gmail.com>",
@@ -30,33 +30,35 @@
     "test": "npm run lint && npm run coverage"
   },
   "peerDependencies": {
-    "@rxstack/core": "^0.2",
-    "@rxstack/exceptions": "^0.2",
-    "@rxstack/platform": "^0.2",
-    "@rxstack/query-filter": "^0.2",
-    "@rxstack/security": "^0.2"
+    "@rxstack/core": "^0.3",
+    "@rxstack/exceptions": "^0.3",
+    "@rxstack/platform": "^0.3",
+    "@rxstack/query-filter": "^0.3",
+    "@rxstack/security": "^0.3"
   },
   "dependencies": {
+    "@types/bluebird": "^3.5.27",
+    "@types/validator": "^10.11.1",
     "injection-js": "^2.1",
     "lodash": "^4.17.11",
     "reflect-metadata": "^0.1",
-    "sequelize": "^4.43.0"
+    "sequelize": "^6.0"
   },
   "devDependencies": {
     "@types/chai": "^4.1.2",
-    "@types/lodash": "^4.14.123",
+    "@types/lodash": "^4.14.135",
     "@types/mocha": "^2.2.48",
-    "@types/node": "^11.12.1",
+    "@types/node": "^11.13.15",
     "chai": "3.5.0",
     "mocha": "^5.2.0",
     "mysql2": "^1.6.5",
-    "nyc": "^13.3.0",
+    "nyc": "^14.1.1",
     "should": "^13.2.3",
     "sinon": "^6.3.4",
     "source-map-support": "^0.4.15",
     "ts-node": "^3.1.0",
-    "tslint": "^5.14.0",
-    "typescript": "^3.3.4000"
+    "tslint": "^5.18.0",
+    "typescript": "^3.5.2"
   },
   "bugs": {
     "url": "https://github.com/rxstack/sequelize-service/issues"
