@@ -1,17 +1,22 @@
 import {ServiceOptions} from '@rxstack/platform';
 import {InjectionToken} from 'injection-js';
 import {LoggingLevel} from '@rxstack/core';
+import {BuildOptions, Model, Options, Sequelize} from 'sequelize';
 
 export interface SequelizeServiceModuleOptions {
-  connection: any;
+  connection: Options;
   logger?: {
     enabled: boolean;
     level?: LoggingLevel;
   };
 }
 
-export const SEQUELIZE_CONNECTION_TOKEN = new InjectionToken<any>('SEQUELIZE_CONNECTION_TOKEN');
+export const SEQUELIZE_CONNECTION_TOKEN = new InjectionToken<Sequelize>('SEQUELIZE_CONNECTION_TOKEN');
+
+export type ModelStatic = typeof Model & {
+  new (values?: object, options?: BuildOptions): Model;
+};
 
 export interface SequelizeServiceOptions extends ServiceOptions {
-  model: any;
+  model: ModelStatic;
 }
