@@ -1,10 +1,12 @@
-process.env.MYSQL_HOST = process.env.MYSQL_HOST || 'mysqldb';
+process.env.MYSQL_HOST = process.env.MYSQL_HOST || 'localhost';
 process.env.MYSQL_DATABASE = process.env.MYSQL_DATABASE || 'test';
 process.env.MYSQL_USER = process.env.MYSQL_USER || 'root';
 process.env.MYSQL_PASSWORD = process.env.MYSQL_PASSWORD === undefined ? 'root' : process.env.MYSQL_PASSWORD;
 
+import {Sequelize} from 'sequelize';
 import {ApplicationOptions} from '@rxstack/core';
 import {
+  ModelStatic,
   SEQUELIZE_CONNECTION_TOKEN,
   SequelizeService, SequelizeServiceModule
 } from '../../src';
@@ -12,7 +14,7 @@ import {InjectionToken} from 'injection-js';
 import {Task} from './task';
 import {defineModels} from './defineModels';
 
-export const MODELS = new InjectionToken<any>('MODELS');
+export const MODELS = new InjectionToken<{[key: string]: ModelStatic}>('MODELS');
 export const TASK_SERVICE = new InjectionToken<SequelizeService<Task>>('TASK_SERVICE');
 export const SEQUELIZE_SERVICE_OPTIONS: ApplicationOptions = {
   imports: [SequelizeServiceModule.configure({
@@ -33,13 +35,13 @@ export const SEQUELIZE_SERVICE_OPTIONS: ApplicationOptions = {
   providers: [
     {
       provide: MODELS,
-      useFactory: (conn: any) => defineModels(conn),
+      useFactory: (conn: Sequelize) => defineModels(conn),
       deps: [SEQUELIZE_CONNECTION_TOKEN],
     },
     {
       provide: TASK_SERVICE,
-      useFactory: (conn: any, models: any) => {
-        return new SequelizeService({ idField: '_id', defaultLimit: 25, model: models['task']});
+      useFactory: (conn: Sequelize, models: {[key: string]: ModelStatic}) => {
+        return new SequelizeService<Task>({ idField: '_id', defaultLimit: 25, model: models['task']});
       },
       deps: [SEQUELIZE_CONNECTION_TOKEN, MODELS],
     }
