@@ -1,10 +1,12 @@
 import {defineTask} from './models/task.model';
 import {defineComment} from './models/comment.model';
+import { Sequelize} from 'sequelize';
+import {ModelStatic} from '../../src';
 
-export const defineModels = (connection: any): Object =>  {
+export const defineModels = (connection: Sequelize): {[key: string]: ModelStatic} =>  {
 
-  const task: any = defineTask(connection);
-  const comment: any = defineComment(connection);
+  const task = defineTask(connection);
+  const comment = defineComment(connection);
 
   comment.belongsTo(task, {
     foreignKey: 'task_id',
